* commit '298f851c20e2ea1d9d0dfffa0be824db412cac2b':
  无jira任务 sonar问题修改
  无jira任务 sonar问题修改

+++ designer-base/src/main/java/com/fr/design/gui/syntax/print/RPrintUtilities.java
@@ -522,8 +522,8 @@ public abstract class RPrintUtilities {
 				return x;
 			int tabSizeInPixels = tabSizeInSpaces * fm.charWidth(' ');
 			int ntabs = (((int) x) - xOffset) / tabSizeInPixels;
-			float reFloat = xOffset + ((ntabs + 1) * tabSizeInPixels);
-			return reFloat;
+			double reFloat = (double) xOffset + (double) ((ntabs + 1) * tabSizeInPixels);
+			return (float) reFloat;
 		}
 
 	}
